+++ test/node/abstract_client.ts
@@ -5,19 +5,19 @@ import { assert } from 'chai'
 import sinon from 'sinon'
 import fs from 'fs'
 import levelStore from 'mqtt-level-store'
-import Store from '../src/lib/store'
+import Store from '../../src/lib/store'
 import serverBuilderFn from './server_helpers_for_client_tests'
-import handlePubrel from '../src/lib/handlers/pubrel'
+import handlePubrel from '../../src/lib/handlers/pubrel'
 import TeardownHelper from './helpers/TeardownHelper'
-import handle from '../src/lib/handlers/index'
-import handlePublish from '../src/lib/handlers/publish'
+import handle from '../../src/lib/handlers/index'
+import handlePublish from '../../src/lib/handlers/publish'
 import mqtt, {
 	IClientOptions,
 	IClientPublishOptions,
 	IClientSubscribeOptions,
 	ISubscriptionMap,
 	ISubscriptionRequest,
-} from '../src'
+} from '../../src'
 import { IPublishPacket, IPubrelPacket, ISubackPacket, QoS } from 'mqtt-packet'
 import { DoneCallback, ErrorWithReasonCode } from 'src/lib/shared'
 import { fail } from 'assert'
@@ -3333,6 +3333,42 @@ export default function abstractTest(server, config, ports) {
 					assert.equal(err.message, 'connack timeout')
 					client.end(true, done)
 				})
+		})
+
+		it('should reconnect on connack error if requested', function _test(t, done) {
+			let connackErrors = 0
+			const rcNotAuthorized = 135
+			const server2 = serverBuilder(config.protocol, (serverClient) => {
+				serverClient.on('connect', () => {
+					const rc = connackErrors === 0 ? rcNotAuthorized : 0
+					const connack =
+						version === 5 ? { reasonCode: rc } : { returnCode: rc }
+					serverClient.connack(connack)
+				})
+			})
+			teardownHelper.addServer(server2)
+			server2.listen(ports.PORTAND50, () => {
+				const client = connect({
+					host: 'localhost',
+					port: ports.PORTAND50,
+					reconnectPeriod: 10,
+					reconnectOnConnackError: true,
+				})
+				teardownHelper.addClient(client)
+				client.on('error', (err) => {
+					assert.instanceOf(err, ErrorWithReasonCode)
+					assert.equal(
+						(err as ErrorWithReasonCode).code,
+						rcNotAuthorized,
+					)
+					assert.equal(connackErrors, 0)
+					connackErrors++
+				})
+				client.on('connect', () => {
+					assert.equal(connackErrors, 1)
+					done()
+				})
+			})
 		})
 
 		it(