+++ test/node/keepaliveManager.ts
@@ -1,5 +1,5 @@
 import { afterEach, beforeEach, describe, it } from 'node:test'
-import KeepaliveManager from '../src/lib/KeepaliveManager'
+import KeepaliveManager from '../../src/lib/KeepaliveManager'
 import { assert } from 'chai'
 import { useFakeTimers, spy, stub } from 'sinon'
 import { MqttClient } from 'src'